feat: rearrange styled components
