+++ src/client/components/Document/index.tsx
@@ -3,7 +3,7 @@ import { routes } from "virtual:routes";
 import { renderRoutes } from "react-router-config";
 import { MDXProvider } from "@mdx-js/react";
 import { ShowCode } from "../../built-in/ShowCode";
-import { StyledMarkdown } from "./styled";
+import { StyledMarkdown, StyledDocument } from "./styled";
 import { ApiTable } from "../../built-in/ApiTable";
 import { useDefaultRoute } from "../../hooks/useDefaultRoute";
 import { ErrorBoundary } from "../ErrorBoundary";
@@ -11,29 +11,31 @@ import { ErrorBoundary } from "../ErrorBoundary";
 const Document = () => {
   useDefaultRoute();
   return (
-    <ErrorBoundary>
-      <StyledMarkdown className="docit-markdown">
-        <MDXProvider
-          components={{
-            ShowCode,
-            ApiTable,
-            Suspense,
-            code: (props) => {
-              return (
-                <code
-                  className={`${[props.className, "docit-code"].join(" ")}`}
-                >
-                  {props.children}
-                </code>
-              );
-            },
-          }}
-        >
-          <Suspense fallback={<></>}>{renderRoutes(routes)}</Suspense>
-        </MDXProvider>
-      </StyledMarkdown>
-    </ErrorBoundary>
+    <StyledDocument className="docit-document">
+      <ErrorBoundary>
+        <StyledMarkdown className="docit-markdown">
+          <MDXProvider
+            components={{
+              ShowCode,
+              ApiTable,
+              Suspense,
+              code: (props) => {
+                return (
+                  <code
+                    className={`${[props.className, "docit-code"].join(" ")}`}
+                  >
+                    {props.children}
+                  </code>
+                );
+              },
+            }}
+          >
+            <Suspense fallback={<></>}>{renderRoutes(routes)}</Suspense>
+          </MDXProvider>
+        </StyledMarkdown>
+      </ErrorBoundary>
+    </StyledDocument>
   );
 };
 
-export default Document;
+export { Document };
